requirements.txt
APScheduler==3.10.4
python-dotenv==1.0.1
anyio==4.12.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
exceptiongroup==1.3.1
google-genai>=1.0.0
google-api-core
google-auth
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Markdown==3.10
pillow==12.0.0
proto-plus
protobuf
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic_core==2.41.5
python-telegram-bot==22.5
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
Babel==2.17.0
cryptography>=41.0.0
aiosqlite>=0.19.0
tenacity>=8.2.0
gTTS>=2.3.0
openai>=1.0.0
anthropic>=0.30.0
